node_module in cli description
There are quite a few errors in this page. I'm just learning, so I can't fix all of them. I went through the npm tutorial too and that seemed fine, except the padding isn't right when I launch the local site. There are references to node_module in the cli, the relative path isn't right, and the sass commands are incorrect.